// src/ts/Enums.ts
// ------------------------------------------------------------------
// Author: Daniel-Percy Wimpff <daniel@wimpff.de>
// Copyright (c) 2007-2025 Daniel-Percy Wimpff <daniel@wimpff.de>, Germany
// MIT License
// ------------------------------------------------------------------

export enum EnumMenuItemType {
	page = "page",
	link = "link",
	menu = "menu"
}

export enum EnumAnimationMode {
	wheel = "wheel",
	shuffle = "shuffle",
	linear = "linear"
}